+++ packages/devtools/src/composite.ts
@@ -1,5 +1,6 @@
 import type { CeramicApi, SignedCommit } from '@ceramicnetwork/common'
 import { Model, type ModelViewsDefinition } from '@ceramicnetwork/stream-model'
+import { StreamID } from '@ceramicnetwork/streamid'
 import type {
   CompositeViewsDefinition,
   EncodedCompositeDefinition,
@@ -188,6 +189,11 @@ export type CreateParams = {
    * Composite schema string.
    */
   schema: string
+  /**
+   * Whether to add the Models to the index or not. If `true` (default), the Ceramic instance
+   * must be authenticated with an admin DID.
+   */
+  index?: boolean
 }
 
 /**
@@ -202,6 +208,11 @@ export type FromJSONParams = {
    * JSON-encoded composite definition.
    */
   definition: EncodedCompositeDefinition
+  /**
+   * Whether to add the Models to the index or not. If `true`, the Ceramic instance must be
+   * authenticated with an admin DID. Defaults to `false`.
+   */
+  index?: boolean
 }
 
 /**
@@ -216,6 +227,11 @@ export type FromModelsParams = CompositeOptions & {
    * Stream IDs of the Models to import in the composite.
    */
   models: Array<string>
+  /**
+   * Whether to add the Models to the index or not. If `true`, the Ceramic instance must be
+   * authenticated with an admin DID. Defaults to `false`.
+   */
+  index?: boolean
 }
 
 /**
@@ -278,7 +294,14 @@ export class Composite {
     )
 
     definition.views = { models: modelsViews }
-    return new Composite({ commits, definition })
+    const composite = new Composite({ commits, definition })
+
+    // By default, add models to the index
+    if (params.index !== false) {
+      await composite.startIndexingOn(params.ceramic)
+    }
+
+    return composite
   }
 
   /**
@@ -299,13 +322,20 @@ export class Composite {
   static async fromJSON(params: FromJSONParams): Promise<Composite> {
     const { models, ...definition } = params.definition
     const commits = decodeSignedMap(models)
-    return new Composite({
+    const composite = new Composite({
       commits,
       definition: {
         ...definition,
         models: await loadModelsFromCommits(params.ceramic, commits),
       },
     })
+
+    // Only add models to the index if explicitly requested
+    if (params.index) {
+      await composite.startIndexingOn(params.ceramic)
+    }
+
+    return composite
   }
 
   /**
@@ -331,8 +361,14 @@ export class Composite {
           .filter(isSignedCommit)
       })
     )
+    const composite = new Composite({ commits, definition })
 
-    return new Composite({ commits, definition })
+    // Only add models to the index if explicitly requested
+    if (params.index) {
+      await composite.startIndexingOn(params.ceramic)
+    }
+
+    return composite
   }
 
   #commits: Record<string, StreamCommits>
@@ -501,6 +537,15 @@ export class Composite {
     const params = this.toParams()
     const definition = setDefinitionViews(toStrictDefinition(params.definition), views, replace)
     return new Composite({ ...params, definition })
+  }
+
+  /**
+   * Configure the Ceramic node to index the models defined in the composite. An authenticated DID
+   * set as admin in the Ceramic node configuration must be attached to the Ceramic instance.
+   */
+  async startIndexingOn(ceramic: CeramicApi): Promise<void> {
+    const modelIDs = Object.keys(this.#definition.models).map(StreamID.fromString)
+    await ceramic.admin.startIndexingModels(modelIDs)
   }
 
   /**