+++ internal-integration-checks/webpack-notes/CHANGELOG.md
@@ -1,5 +1,14 @@
 # webpack-notes-app
 
+## 0.3.0-rc.1
+
+### Patch Changes
+
+- Add admin APIs support for composites creation and deployment
+- Updated dependencies
+  - @composedb/client@0.3.0-rc.1
+  - @composedb/types@0.3.0-rc.1
+
 ## 0.3.0-rc.0
 
 ### Minor Changes